Ignore flakey log-cache-acceptance-test
log-cache-acceptance-tests run the upstream acceptance tests with a flag
set to make them use `cf tail` instead of `cf logs`. Some of the tests
make assertions about the log output which are flakey when using
logcache like this (e.g. the order of log lines is not guaranteed).

The `applies correct environment variables while running apps` has
failed a few times recently and then passed on subsequent runs with no
changes, e.g.:

https://deployer.london.staging.cloudpipeline.digital/teams/main/pipelines/create-cloudfoundry/jobs/log-cache-acceptance-tests/builds/180
https://deployer.london.staging.cloudpipeline.digital/teams/main/pipelines/create-cloudfoundry/jobs/log-cache-acceptance-tests/builds/176
https://deployer.london.staging.cloudpipeline.digital/teams/main/pipelines/create-cloudfoundry/jobs/log-cache-acceptance-tests/builds/164

(Note: I'm leaving the `a` off `applies` deliberately in the regex
because I'm not sure whether it should be `a` or `A` and I don't care
enough to find out.)