Migrate alert "total routes drop" to Prometheus
Original datadog monitor:

pct_change(avg(last_1m),last_30m):avg:cf.gorouter.total_routes < -33

from terraform/datadog/router.tf total_routes_drop

In this case we add a record to keep the delta in a 5 minute window,
and then we keep the minimun within the last 30m to ensure we alert
for at least 30m

The percentage is computed comparing with the mean in a 5m window.

min_over_time can only work with records, and not with subqueries.
This is a limitation by design in prometheus[1]

[1] https://www.robustperception.io/composing-range-vector-functions-in-promql