Run rotate_cc_database_key errand in post-deploy
We want to be able to run the rotate_cc_database_key errand job
to rotate the DB keys from the CC if they changed[1]

We will run it only on the first api VM in the post-deploy job,
adding the `--when-changed` to run it only if the config changed.
That is mainly if the keys changed.

It could be triggered due other config changes, but it is an
idempotent operation and that is fine.

For the record, we tried change this to run as a colocated job in
the api nodes, so that we do not need to create an adhoc VM with
permissions to access the DB,  and it would be easier to troubleshoot
or do manual interventions. But this required to grant access
to the blobstore to the VMs API and UAA[2], which requires multiple
changes

[1] https://github.com/cloudfoundry/capi-release/tree/develop/jobs/rotate_cc_database_key
[2] https://github.com/cloudfoundry/bosh/issues/1150
[3] https://github.com/alphagov/paas-aws-account-wide-terraform/pull/132/files