Add 9 more cells in London
We are currently well below the alarm threshold in London for the available memory on cells. This means we need to add more cells.

`rep_memory_capacity_pct:avg5m` is sitting around 22.8%, which is less than the threshold of 33.

In the past, we have used:

```
$current_memory_use + $current_memory_use * $num_cells_to_add/$existing_cells
```

Therefore, we should scale up by 9 cells to 27: `22.8 + 22.8 * 9/18 = 34.2`