Temporarily kick off dev environments every day
Concourse 5 replaced Resource Pausing with Pinning. Pinning does not
work quite the same way, and it seems to have broken the way our dev
environments set themselves up each morning.

This commit makes dev environments kick-off every morning, without
having to opt-in. This will cost lots of money if the root problem is
not dealt with.