Make a highly-available dev environment
In Pivotal story #166056703 we are rotating the certificates used by
our production environments. We have not done this in some time and
want to be sure the documented process [1] still works correctly and
without downtime.

This commit changes any dev environment deployed from it to have more
instances of various VMs so that it is highly-available and is suited
for testing that certificate rotation is zero-downtime and works well.

[1] https://team-manual.cloud.service.gov.uk/team/rotating_credentials/#rotating-certificates