WIP: add monitor for version of github releases
The idea here is that we'd report the latest and current versions of
our dependencies to prometheus. We could then track how many versions we
are behind.

Mainly this would be useful for cf-deployment, but could also be used
for things like bosh and the prometheus boshrelease.

This commit adds *untested* code to get the latest version of things
that have github releases.

We could get the version from master using some other github apis.

We may need to think about giving paas metrics a github token so it
doesn't get rate limited (github will only see the IP of our NAT
gateways, so everything on the PaaS will share the same rate limit if
they don't use tokens).