Replace Doppler ELB with NLB
Loggregator release 105.6 changes the TLS ciphers used by Doppler, and the new
ones aren't supported by classic load balancers. In order to upgrade in the
future, we're replacing the ELB and an NLB.

In this commit, we introduce the new NLB, point the DNS records at it, and have
the Doppler instances register themselves in its target group. A future commit
will remove the ELB.